+++ planning/autoware_minimum_rule_based_planner/CMakeLists.txt
@@ -16,7 +16,6 @@ ament_auto_add_library(${PROJECT_NAME}_lib SHARED
   src/map_based_stop_planner.cpp
   src/path_planner.cpp
   src/turn_indicator_decider.cpp
-  src/utils/turn_indicator_utils.cpp
   src/velocity_smoother.cpp
 )
