+++ planning/autoware_minimum_rule_based_planner/config/minimum_rule_based_planner.param.yaml
@@ -176,4 +176,3 @@
       enable_optimizer_trajectory: true
       enable_modifier_trajectory: true
       enable_output_trajectory: true
-      enable_turn_indicator_marker: true
