+++ planning/autoware_minimum_rule_based_planner/src/turn_indicator_decider.cpp
@@ -21,10 +21,6 @@
 #include <autoware/motion_utils/trajectory/trajectory.hpp>
 #include <autoware_utils_geometry/geometry.hpp>
 #include <autoware_utils_math/normalization.hpp>
-#include <magic_enum.hpp>
-#include <rclcpp/duration.hpp>
-
-#include <std_msgs/msg/color_rgba.hpp>
 
 #include <tf2/utils.h>
 
@@ -32,6 +28,7 @@
 #include <cmath>
 #include <cstddef>
 #include <cstdint>
+#include <initializer_list>
 #include <limits>
 #include <optional>
 #include <string>
@@ -39,6 +36,116 @@
 
 namespace autoware::minimum_rule_based_planner
 {
+
+namespace turn_indicator
+{
+
+double activation_distance(const double ego_velocity, const TurnSignalParams & params)
+{
+  return std::max(ego_velocity * k_search_time, params.search_distance);
+}
+
+TurnDirection direction_from_lateral_offset(const double signed_offset, const double deadzone)
+{
+  if (signed_offset > deadzone) {
+    return TurnDirection::LEFT;
+  }
+  if (signed_offset < -deadzone) {
+    return TurnDirection::RIGHT;
+  }
+  return TurnDirection::NONE;
+}
+
+Signal resolve_priority(const std::initializer_list<Signal> candidates)
+{
+  for (const auto & candidate : candidates) {
+    if (candidate.direction != TurnDirection::NONE) {
+      return candidate;
+    }
+  }
+  return {};
+}
+
+TurnDirection decide_maneuver_signal(
+  const TurnDirection direction, const double dist_to_start, const double ego_yaw,
+  const double exit_yaw, const double ego_velocity, const TurnSignalParams & params)
+{
+  if (direction == TurnDirection::NONE) {
+    return TurnDirection::NONE;
+  }
+
+  constexpr double entered_tolerance = 1e-3;
+  const double yaw_gap = std::abs(autoware_utils_math::normalize_radian(ego_yaw - exit_yaw));
+  if (dist_to_start <= entered_tolerance && yaw_gap <= params.heading_align_threshold) {
+    return TurnDirection::NONE;
+  }
+  if (dist_to_start <= activation_distance(ego_velocity, params)) {
+    return direction;
+  }
+  return TurnDirection::NONE;
+}
+
+TurnDirection decide_pull_out(
+  const double signed_offset, const double ego_velocity, const bool suppressed,
+  const TurnSignalParams & params, TurnDirection & latched)
+{
+  const double offset_magnitude = std::abs(signed_offset);
+
+  if (suppressed || offset_magnitude <= k_lateral_shift_threshold) {
+    latched = TurnDirection::NONE;
+    return TurnDirection::NONE;
+  }
+
+  if (latched == TurnDirection::NONE) {
+    const bool stopped = std::abs(ego_velocity) <= params.stopped_velocity_threshold;
+    if (!stopped || offset_magnitude <= k_departure_lateral_threshold) {
+      return TurnDirection::NONE;
+    }
+    latched = direction_from_lateral_offset(-signed_offset, k_lateral_shift_threshold);
+  }
+
+  return latched;
+}
+
+TurnDirection decide_pull_over(
+  const double dist_to_goal, const double goal_offset, const double ego_velocity,
+  const TurnSignalParams & params, bool & arrived)
+{
+  if (dist_to_goal > params.search_distance) {
+    arrived = false;
+    return TurnDirection::NONE;
+  }
+
+  if (
+    dist_to_goal <= k_goal_arrival_distance &&
+    std::abs(ego_velocity) <= params.stopped_velocity_threshold) {
+    arrived = true;
+  }
+  if (arrived) {
+    return TurnDirection::NONE;
+  }
+
+  return direction_from_lateral_offset(goal_offset, k_lateral_shift_threshold);
+}
+
+TurnDirection BlinkHold::update(const TurnDirection desired, const double now)
+{
+  if (desired == held_) {
+    return held_;
+  }
+
+  const bool turning_off = desired == TurnDirection::NONE;
+  if (turning_off && held_ != TurnDirection::NONE && (now - held_since_) < min_duration_) {
+    return held_;
+  }
+
+  held_ = desired;
+  held_since_ = now;
+  return held_;
+}
+
+}  // namespace turn_indicator
+
 namespace
 {
 namespace lanelet2_utils = autoware::experimental::lanelet2_utils;
@@ -59,19 +166,12 @@ uint8_t to_command(const TurnDirection direction)
   }
 }
 
-std::string to_string(const Signal & signal)
-{
-  return std::string(magic_enum::enum_name(signal.direction)) + " (" +
-         std::string(magic_enum::enum_name(signal.kind)) + ")";
-}
-
 bool is_private(const lanelet::ConstLanelet & lanelet)
 {
   return lanelet.attributeOr(lanelet::AttributeNamesString::Location, std::string("")) ==
          lanelet::AttributeValueString::Private;
 }
 
-//! The lanelet with `id`, or nothing when the map does not hold it.
 std::optional<lanelet::ConstLanelet> find_lanelet(const lanelet::LaneletMap & map, const int64_t id)
 {
   try {
@@ -93,8 +193,7 @@ TurnDirection tagged_turn_direction(const lanelet::ConstLanelet & lanelet)
   return TurnDirection::NONE;
 }
 
-//! Path heading `offset` metres along the path from `index` (signed), falling back to the path end
-//! (or start) when the path is too short to reach it.
+//! Path heading `offset` metres (signed) along the path from `index`, clamped to the path ends.
 double path_yaw_at(const PathWithLaneId & path, const std::size_t index, const double offset)
 {
   const auto pose = motion_utils::calcLongitudinalOffsetPose(path.points, index, offset);
@@ -105,17 +204,14 @@ double path_yaw_at(const PathWithLaneId & path, const std::size_t index, const d
   return tf2::getYaw(fallback.point.pose.orientation);
 }
 
-//! An upcoming (or currently occupied) maneuver along the path.
 struct Maneuver
 {
   Signal signal;
   double dist_to_start{0.0};  //!< [m] arc length from ego to its start (negative once inside)
   double exit_yaw{0.0};       //!< [rad] path heading the maneuver ends on
-  std::size_t start_index{0};
-  std::size_t end_index{0};
 };
 
-//! Distinct lane ids in the order they first appear along the path.
+//! Distinct lane ids in first-appearance order.
 std::vector<int64_t> ordered_lane_ids(const PathWithLaneId & path)
 {
   std::vector<int64_t> ids;
@@ -129,15 +225,11 @@ std::vector<int64_t> ordered_lane_ids(const PathWithLaneId & path)
   return ids;
 }
 
-//! Every maneuver this module signals for, ordered from ego. Two cases, both keyed on the lanelet
-//! run the path traverses:
+//! Maneuvers along the path, ordered from ego. Two cases:
 //!  - an intersection turn: a lanelet tagged `turn_direction=left/right`;
 //!  - a private-area exit: the boundary where a `location=private` lanelet rejoins a public one.
-//! The private case is handled separately because such a lanelet may be tagged `straight`, or not
-//! tagged at all, so the turn tag alone would miss it.
-//!
-//! A maneuver whose exit lies more than `k_exit_lookahead` behind ego is dropped; one ego is still
-//! completing is kept, so its signal survives until the heading check clears it.
+//! The private case needs its own detection because such a lanelet is usually untagged in practice
+//! (or tagged `straight`), so the turn tag alone would miss the merge.
 std::vector<Maneuver> find_maneuvers(
   const PathWithLaneId & path, const std::size_t ego_index, const RouteContext & route_context,
   const turn_indicator::TurnSignalParams & params)
@@ -159,9 +251,8 @@ std::vector<Maneuver> find_maneuvers(
       continue;
     }
 
-    // The private run ends here only if the next lanelet along the path is public (or the path
-    // ends). A run of several private lanelets therefore yields exactly one exit. An id the map
-    // does not hold counts as public, so a missing lanelet cannot swallow the exit.
+    // A run of several private lanelets yields exactly one exit: the last one before a public
+    // lanelet (an id the map does not hold counts as public, so it cannot swallow the exit).
     const auto next = i + 1 < lane_ids.size() ? find_lanelet(*map, lane_ids[i + 1]) : std::nullopt;
     const bool leaves_private_area = is_private(*lanelet) && (!next || !is_private(*next));
     const auto turn_direction = tagged_turn_direction(*lanelet);
@@ -169,21 +260,20 @@ std::vector<Maneuver> find_maneuvers(
       continue;
     }
 
+    // A turn starts at the entry of its lanelet; a private exit at the boundary itself (the
+    // private run leading up to it may be long).
+    const std::size_t start_index =
+      turn_direction != TurnDirection::NONE ? range->first : range->second;
+
     Maneuver maneuver;
-    // A turn is signalled from the entry of its lanelet; a private exit from the boundary itself,
-    // which is where the maneuver actually starts (the private run leading up to it may be long).
-    maneuver.start_index = turn_direction != TurnDirection::NONE ? range->first : range->second;
-    maneuver.end_index = range->second;
     maneuver.exit_yaw = path_yaw_at(path, range->second, turn_indicator::k_exit_lookahead);
-    maneuver.dist_to_start =
-      motion_utils::calcSignedArcLength(path.points, ego_index, maneuver.start_index);
+    maneuver.dist_to_start = motion_utils::calcSignedArcLength(path.points, ego_index, start_index);
 
     if (turn_direction != TurnDirection::NONE) {
-      maneuver.signal = {turn_direction, ManeuverKind::INTERSECTION};
+      maneuver.signal = {turn_direction, ManeuverKind::TURN};
     } else {
-      // Untagged private exit: take the side from the yaw change across the merge, measured from a
-      // look-ahead BEFORE the boundary - right at it the path is already turning, so the local yaw
-      // change is near zero and every merge would read as straight.
+      // Untagged private exit: the side comes from the yaw change across the merge, measured
+      // from before the boundary - at it the path is already turning, so it reads as straight.
       const double approach_yaw =
         path_yaw_at(path, range->second, -turn_indicator::k_exit_lookahead);
       const double delta = autoware_utils_math::normalize_radian(maneuver.exit_yaw - approach_yaw);
@@ -191,11 +281,11 @@ std::vector<Maneuver> find_maneuvers(
         continue;  // a geometrically straight merge has no side to signal
       }
       maneuver.signal = {
-        delta > 0.0 ? TurnDirection::LEFT : TurnDirection::RIGHT, ManeuverKind::PRIVATE_EXIT};
+        delta > 0.0 ? TurnDirection::LEFT : TurnDirection::RIGHT, ManeuverKind::TURN};
     }
 
     const double dist_to_end =
-      motion_utils::calcSignedArcLength(path.points, ego_index, maneuver.end_index);
+      motion_utils::calcSignedArcLength(path.points, ego_index, range->second);
     if (dist_to_end + turn_indicator::k_exit_lookahead <= 0.0) {
       continue;  // fully behind ego
     }
@@ -204,8 +294,8 @@ std::vector<Maneuver> find_maneuvers(
   return maneuvers;
 }
 
-//! The lane ego is expected to track plus its lateral neighbours, so that a lane change (whose
-//! target lane is not the path's lane) still reports a near-zero offset and cannot raise a signal.
+//! The lane ego tracks plus its lateral neighbours, so a lane change (whose target lane is not the
+//! path's lane) still reports a near-zero offset and cannot raise a signal.
 lanelet::ConstLanelets ego_lanes(
   const PathWithLaneId & path, const std::size_t ego_index, const RouteContext & route_context)
 {
@@ -241,17 +331,15 @@ TurnIndicatorsCommand TurnIndicatorDecider::decide(
   const PathWithLaneId & path, const RouteContext & route_context,
   const geometry_msgs::msg::Pose & ego_pose, const double ego_velocity, const rclcpp::Time & stamp)
 {
-  debug_ = TurnIndicatorDebug();
   TurnIndicatorsCommand cmd;
   cmd.stamp = stamp;
 
   if (path.points.empty()) {
     cmd.command = to_command(blink_hold_.update(TurnDirection::NONE, stamp.seconds()));
-    debug_.signal.direction = blink_hold_.current();
     return cmd;
   }
 
-  // Re-arm the latched pull-out / pull-over states whenever the route changes.
+  // Re-arm the latched pull-out / pull-over states on a route change.
   if (
     !latched_goal_pose_ ||
     autoware_utils_geometry::calc_distance2d(*latched_goal_pose_, route_context.goal_pose) > 1e-3) {
@@ -260,34 +348,27 @@ TurnIndicatorsCommand TurnIndicatorDecider::decide(
     arrived_at_goal_ = false;
   }
 
-  // Reject path points the path traverses in (roughly) the opposite direction to ego, so a route
-  // folding back on itself (U-turn, loop) cannot snap ego onto the wrong leg.
+  // The yaw limit rejects points facing away from ego, so a route folding back on itself (U-turn,
+  // loop) cannot snap ego onto the wrong leg.
   const double ego_yaw = tf2::getYaw(ego_pose.orientation);
   const auto aligned_index = motion_utils::findNearestIndex(
     path.points, ego_pose, std::numeric_limits<double>::max(), M_PI_2);
   const std::size_t ego_index =
     aligned_index.value_or(motion_utils::findNearestIndex(path.points, ego_pose.position));
 
-  // 1./2. Intersection turns and private-area exits, both keyed on the path's lanelets. The
-  //       nearest lit maneuver of each kind competes; an intersection outranks a private exit.
-  const Maneuver * lit_intersection = nullptr;
-  const Maneuver * lit_private_exit = nullptr;
-  const auto maneuvers = find_maneuvers(path, ego_index, route_context, params_);
-  for (const auto & maneuver : maneuvers) {
+  // Maneuvers come back in path order, so the first one asking for a light is the one ego reaches
+  // first - or is still completing, which keeps its signal from being stolen by the next one.
+  Signal maneuver_signal;
+  for (const auto & maneuver : find_maneuvers(path, ego_index, route_context, params_)) {
     const auto direction = turn_indicator::decide_maneuver_signal(
       maneuver.signal.direction, maneuver.dist_to_start, ego_yaw, maneuver.exit_yaw, ego_velocity,
       params_);
-    if (direction == TurnDirection::NONE) {
-      continue;
-    }
-    const Maneuver *& slot =
-      maneuver.signal.kind == ManeuverKind::INTERSECTION ? lit_intersection : lit_private_exit;
-    if (slot == nullptr) {
-      slot = &maneuver;
+    if (direction != TurnDirection::NONE) {
+      maneuver_signal = maneuver.signal;
+      break;
     }
   }
 
-  // 3. Arrival (pull-over): signal toward the side an off-centerline goal sits on.
   const double dist_to_goal =
     autoware_utils_geometry::calc_distance2d(ego_pose, route_context.goal_pose);
   const double goal_offset = route_context.goal_lanelets.empty()
@@ -297,8 +378,7 @@ TurnIndicatorsCommand TurnIndicatorDecider::decide(
   const auto pull_over = turn_indicator::decide_pull_over(
     dist_to_goal, goal_offset, ego_velocity, params_, arrived_at_goal_);
 
-  // 4. Departure (pull-out). Suppressed inside the pull-over range so the two cannot fight over
-  //    the direction while ego pulls into an offset goal.
+  // Pull-out is suppressed inside the pull-over range so the two cannot fight over the direction.
   const auto lanes = ego_lanes(path, ego_index, route_context);
   const auto pull_out =
     lanes.empty()
@@ -309,64 +389,11 @@ TurnIndicatorsCommand TurnIndicatorDecider::decide(
           dist_to_goal <= params_.search_distance, params_, pull_out_latch_);
 
   const auto signal = turn_indicator::resolve_priority(
-    {lit_intersection != nullptr ? lit_intersection->signal : Signal{},
-     lit_private_exit != nullptr ? lit_private_exit->signal : Signal{},
-     Signal{pull_out, ManeuverKind::PULL_OUT}, Signal{pull_over, ManeuverKind::PULL_OVER}});
+    {maneuver_signal, Signal{pull_out, ManeuverKind::PULL_OUT},
+     Signal{pull_over, ManeuverKind::PULL_OVER}});
   cmd.command = to_command(blink_hold_.update(signal.direction, stamp.seconds()));
 
-  // Record what drove the decision, for visualization.
-  debug_.signal = {blink_hold_.current(), signal.kind};
-  const Maneuver * active = signal.kind == ManeuverKind::INTERSECTION   ? lit_intersection
-                            : signal.kind == ManeuverKind::PRIVATE_EXIT ? lit_private_exit
-                                                                        : nullptr;
-  if (active != nullptr) {
-    debug_.has_segment = true;
-    debug_.start_point = path.points.at(active->start_index).point.pose.position;
-    debug_.end_point = path.points.at(active->end_index).point.pose.position;
-  }
-
   return cmd;
-}
-
-visualization_msgs::msg::MarkerArray create_turn_indicator_markers(
-  const TurnIndicatorDebug & debug, const std_msgs::msg::Header & header)
-{
-  visualization_msgs::msg::MarkerArray markers;
-
-  const bool lit = debug.signal.direction != turn_indicator::TurnDirection::NONE;
-  std_msgs::msg::ColorRGBA colour;  // amber when lit, grey when dark
-  colour.a = 0.999;
-  colour.r = lit ? 1.0 : 0.6;
-  colour.g = lit ? 0.7 : 0.6;
-  colour.b = lit ? 0.0 : 0.6;
-
-  visualization_msgs::msg::Marker marker;
-  marker.header = header;
-  marker.action = visualization_msgs::msg::Marker::ADD;
-  marker.lifetime = rclcpp::Duration::from_seconds(0.5);
-  marker.pose.orientation.w = 1.0;
-  marker.color = colour;
-
-  if (debug.has_segment) {
-    marker.ns = "turn_indicator_segment";
-    marker.type = visualization_msgs::msg::Marker::ARROW;
-    marker.scale.x = 0.4;  // shaft diameter
-    marker.scale.y = 0.8;  // head diameter
-    marker.points = {debug.start_point, debug.end_point};
-    markers.markers.push_back(marker);
-    marker.points.clear();
-  }
-
-  marker.ns = "turn_indicator_state";
-  marker.type = visualization_msgs::msg::Marker::TEXT_VIEW_FACING;
-  marker.pose.position = debug.start_point;
-  marker.pose.position.z += 2.0;
-  marker.scale = geometry_msgs::msg::Vector3();
-  marker.scale.z = 1.0;
-  marker.text = to_string(debug.signal);
-  markers.markers.push_back(marker);
-
-  return markers;
 }
 
 }  // namespace autoware::minimum_rule_based_planner
